requirements.txt
aiofiles
aiohttp>=3.9.1
biopython
cohere
deep-translator
docx2pdf
easyocr
fastapi
fastembed
frontend
httpx-oauth
icecream
instructor
IPython
llama-hub
llama-index-core
llama-index-embeddings-fastembed
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-retrievers-bm25
llama-index-indices-managed-vectara
llama-index-vector-stores-qdrant
llama-index-vector-stores-pinecone
llama-index-readers-box
box-sdk-gen
llama-parse
llmsherpa
nest_asyncio
numpy
onnxruntime
openai
openpyxl
optimum
optimum[exporters]
optimum[onnxruntime]
pandas
pdfminer
portalocker
psutil
pydantic
python-dotenv>=1.0.0
pyvis
markdown>=3.5.1
qdrant_client
qdrant-client[fastembed]
rank-bm25
reportlab
scikit-learn
sentence-transformers
spider-client
streamlit>=1.24.0
streamlit-antd-components
streamlit-authenticator
streamlit-camera-input-live
streamlit-card
streamlit-carousel
streamlit-elements
streamlit-extras
streamlit-image-select
streamlit-searchbox
streamlit-webrtc
st_login_form
transformers
unstructured
unstructured[docx]
unstructured[pdf]
unstructured_inference
uvicorn
xlsxwriter
onnx
reportlab
plotly
networkx
beautifulsoup4>=4.12.2
requests>=2.31.0
pydantic-ai[logfire]
selenium>=4.15.2
webdriver-manager>=4.0.1
tavily-python
PyMuPDF>=1.23.7